refactor: migrate scroll model from line-count to pixel-space (f32)
Change scroll_offset and all scroll parameters from integer line counts
to f32 pixel offsets. This is a pure refactor — behavior is identical
(scroll still jumps by whole lines) but the internal representation now
uses pixels, which enables smooth scrolling in a follow-up task.

Key changes:
- CoreState.scroll_offset: u32 → f32 (pixel offset, not line count)
- CoreState.saved_editor_scroll: u32 → f32
- scroll_for_cursor(): returns f32 pixel offset instead of line index
- scroll_runs(): accepts scroll_y: f32 (pixel space) instead of
  scroll_lines: u32
- All layout functions: scroll_y: i32 → scroll_y: f32
- content_transform uses scroll_y directly (no intermediate conversion)
- Click hit-testing uses round_f32() instead of line * height
- Scroll comparison uses half-pixel epsilon instead of integer equality
- Added round_f32() helper for no_std (f32::round() unavailable)

+++ system/services/core/scene_state.rs
@@ -51,7 +51,7 @@ impl SceneState {
         sel_end: u32,
         title_label: &[u8],
         clock_text: &[u8],
-        scroll_y: i32,
+        scroll_y: f32,
     ) {
         let mut tw = self.triple();
         {
@@ -140,7 +140,7 @@ impl SceneState {
         changed_line: usize,
         title_label: &[u8],
         clock_text: &[u8],
-        scroll_y: i32,
+        scroll_y: f32,
         timer_fired: bool,
     ) {
         let mut tw = self.triple();
@@ -188,7 +188,7 @@ impl SceneState {
         sel_end: u32,
         title_label: &[u8],
         clock_text: &[u8],
-        scroll_y: i32,
+        scroll_y: f32,
         timer_fired: bool,
     ) {
         let mut tw = self.triple();
@@ -234,7 +234,7 @@ impl SceneState {
         sel_end: u32,
         title_label: &[u8],
         clock_text: &[u8],
-        scroll_y: i32,
+        scroll_y: f32,
         timer_fired: bool,
     ) {
         let mut tw = self.triple();
@@ -280,7 +280,7 @@ impl SceneState {
         sel_end: u32,
         title_label: &[u8],
         clock_text: &[u8],
-        scroll_y: i32,
+        scroll_y: f32,
         mark_clock_changed: bool,
     ) {
         let mut tw = self.triple();
